+++ src/main/java/gregtech/api/world/GT_Worldgen.java
@@ -2,8 +2,8 @@ package gregtech.api.world;
 
 import com.google.common.base.Predicate;
 import gregtech.api.GregTech_API;
-import gregtech.common.blocks.GT_Block_Granites;
-import gregtech.common.blocks.GT_Block_Stones;
+import gregtech.common.blocks.BlockGranites;
+import gregtech.common.blocks.BlockStones;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.math.BlockPos;
@@ -70,8 +70,8 @@ public abstract class GT_Worldgen {
 
     public static Predicate<IBlockState> STONES = input ->
             input.getBlock() == Blocks.STONE ||
-                    input.getBlock() instanceof GT_Block_Granites ||
-                    input.getBlock() instanceof GT_Block_Stones;
+                    input.getBlock() instanceof BlockGranites ||
+                    input.getBlock() instanceof BlockStones;
 
     public static Predicate<IBlockState> NETHERRACK = input ->
             input.getBlock() == Blocks.NETHERRACK;
