+++ src/main/java/gregtech/common/GT_Worldgen_Stone.java
@@ -3,6 +3,7 @@ package gregtech.common;
 import gregtech.api.enums.StoneTypes;
 import gregtech.api.world.GT_Worldgen;
 import gregtech.api.world.GT_Worldgen_Ore;
+import gregtech.common.blocks.BlockGeneratedOres;
 import gregtech.common.blocks.GT_Block_GeneratedOres;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.math.BlockPos;
@@ -68,8 +69,8 @@ public class GT_Worldgen_Stone
                                             if (var12d + var45 * var45 < 1.0F) {
                                                 temp.setPos(eX, eY, eZ);
                                                 IBlockState tTargetedBlock = aWorld.getBlockState(temp);
-                                                if (tTargetedBlock.getBlock() instanceof GT_Block_GeneratedOres) {
-                                                    GT_Block_GeneratedOres block = (GT_Block_GeneratedOres) tTargetedBlock.getBlock();
+                                                if (tTargetedBlock.getBlock() instanceof BlockGeneratedOres) {
+                                                    BlockGeneratedOres block = (BlockGeneratedOres) tTargetedBlock.getBlock();
                                                     StoneTypes stoneTypes = StoneTypes.computeStoneType(mBlockState);
                                                     aWorld.setBlockState(temp, block.overrideStoneType(tTargetedBlock, stoneTypes));
                                                 } else if ((this.mAllowToGenerateinVoid && aWorld.isAirBlock(temp)) ||
