+++ src/main/java/gregtech/common/blocks/itemblocks/ItemCasings1.java
@@ -2,8 +2,8 @@ package gregtech.common.blocks.itemblocks;
 
 import net.minecraft.block.Block;
 
-public class GT_Item_Casings1 extends GT_Item_Casings_Abstract {
-    public GT_Item_Casings1(Block block) {
+public class ItemCasings1 extends ItemCasingsAbstract {
+    public ItemCasings1(Block block) {
         super(block);
     }
 }
