+++ src/main/java/gregtech/common/blocks/itemblocks/ItemCasings3.java
@@ -2,8 +2,8 @@ package gregtech.common.blocks.itemblocks;
 
 import net.minecraft.block.Block;
 
-public class GT_Item_Casings4 extends GT_Item_Casings_Abstract {
-    public GT_Item_Casings4(Block block) {
+public class ItemCasings3 extends ItemCasingsAbstract {
+    public ItemCasings3(Block block) {
         super(block);
     }
 }
