+++ src/main/java/gregtech/common/blocks/itemblocks/ItemConcretes.java
@@ -7,16 +7,16 @@ import net.minecraft.item.ItemStack;
 
 import java.util.List;
 
-public class GT_Item_Concretes extends GT_Item_Stones_Abstract {
-    private final String mRunFasterToolTip = GT_LanguageManager.addStringLocalization("gt.runfastertooltip", "You can walk faster on this Block");
+public class ItemConcretes extends ItemStonesAbstract {
+    private final String runFasterToolTip = GT_LanguageManager.addStringLocalization("gt.runfastertooltip", "You can walk faster on this Block");
 
-    public GT_Item_Concretes(Block block) {
+    public ItemConcretes(Block block) {
         super(block);
     }
 
     @Override
     public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean advanced) {
         super.addInformation(stack, player, tooltip, advanced);
-        tooltip.add(this.mRunFasterToolTip);
+        tooltip.add(this.runFasterToolTip);
     }
 }
