+++ src/main/java/gregtech/common/blocks/itemblocks/ItemGranites.java
@@ -2,8 +2,8 @@ package gregtech.common.blocks.itemblocks;
 
 import net.minecraft.block.Block;
 
-public class GT_Item_Casings3 extends GT_Item_Casings_Abstract {
-    public GT_Item_Casings3(Block block) {
+public class ItemGranites extends ItemStonesAbstract {
+    public ItemGranites(Block block) {
         super(block);
     }
 }
