+++ src/main/java/gregtech/common/blocks/itemblocks/ItemStonesAbstract.java
@@ -9,11 +9,11 @@ import net.minecraft.item.ItemStack;
 
 import java.util.List;
 
-public class GT_Item_Stones_Abstract extends ItemBlock {
+public class ItemStonesAbstract extends ItemBlock {
 
-    private final String mNoMobsToolTip = GT_LanguageManager.addStringLocalization("gt.nomobspawnsonthisblock", "Mobs cannot Spawn on this Block");
+    private final String noMobsToolTip = GT_LanguageManager.addStringLocalization("gt.nomobspawnsonthisblock", "Mobs cannot Spawn on this Block");
 
-    public GT_Item_Stones_Abstract(Block block) {
+    public ItemStonesAbstract(Block block) {
         super(block);
         setMaxDamage(0);
         setHasSubtypes(true);
@@ -39,7 +39,7 @@ public class GT_Item_Stones_Abstract extends ItemBlock {
     public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean advanced) {
         super.addInformation(stack, player, tooltip, advanced);
         if (stack.getItemDamage() % 8 >= 3) {
-            tooltip.add(this.mNoMobsToolTip);
+            tooltip.add(this.noMobsToolTip);
         }
     }
 }
