+++ src/main/java/gregtech/common/blocks/materials/MaterialCasings.java
@@ -3,14 +3,13 @@ package gregtech.common.blocks.materials;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 
-public class GT_Material_Reinforced extends Material {
+public class MaterialCasings extends Material {
 
-    public static final Material INSTANCE = new GT_Material_Reinforced();
+    public static final Material INSTANCE = new MaterialCasings();
 
-    public GT_Material_Reinforced() {
-        super(MapColor.STONE);
+    public MaterialCasings() {
+        super(MapColor.IRON);
         setRequiresTool();
-        setAdventureModeExempt();
     }
 
     @Override
