+++ src/main/java/gregtech/common/blocks/materials/MaterialMachines.java
@@ -3,11 +3,11 @@ package gregtech.common.blocks.materials;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 
-public class GT_Material_Machines extends Material {
+public class MaterialMachines extends Material {
 
-    public static final Material INSTANCE = new GT_Material_Machines();
+    public static final Material INSTANCE = new MaterialMachines();
 
-    public GT_Material_Machines() {
+    public MaterialMachines() {
         super(MapColor.IRON);
         setRequiresTool();
         setImmovableMobility();
