+++ src/main/java/gregtech/common/blocks/materials/MaterialReinforced.java
@@ -3,13 +3,14 @@ package gregtech.common.blocks.materials;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 
-public class GT_Material_Casings extends Material {
+public class MaterialReinforced extends Material {
 
-    public static final Material INSTANCE = new GT_Material_Casings();
+    public static final Material INSTANCE = new MaterialReinforced();
 
-    public GT_Material_Casings() {
-        super(MapColor.IRON);
+    public MaterialReinforced() {
+        super(MapColor.STONE);
         setRequiresTool();
+        setAdventureModeExempt();
     }
 
     @Override
